Revert "added loggedin navbar"
This reverts commit 165fda0eeb6e67adfce9cd349a096f9718fcfd3a.
