+++ client/.eslintrc
@@ -1,23 +1,23 @@
 {
   "parser": "@typescript-eslint/parser",
   "parserOptions": {
-    "ecmaVersion": 2021,
-    "sourceType": "module"
+      "ecmaVersion": 2021,
+      "sourceType": "module"
   },
   "extends": [
-    "plugin:@typescript-eslint/recommended",
-    "plugin:react/recommended",
-    //   "plugin:prettier/recommended",
-    "plugin:react-hooks/recommended"
+      "plugin:@typescript-eslint/recommended",
+      "plugin:react/recommended",
+      "plugin:prettier/recommended",
+      "plugin:react-hooks/recommended"
   ],
   "settings": {
-    "react": { "version": "detect" }
+    "react": { "version": "detect"}
   },
   "rules": {
     // suppress errors for missing 'import React' in files
-    "react/react-in-jsx-scope": "off",
-    "react/prop-types": "off",
-    "react-hooks/rules-of-hooks": "error",
-    "react-hooks/exhaustive-deps": "error"
+   "react/react-in-jsx-scope": "off",
+   "react/prop-types": "off",
+   "react-hooks/rules-of-hooks": "error",
+   "react-hooks/exhaustive-deps": "error"
   }
 }
