--- client/src/components/AuthMenu/useStyles.ts
@@ -1,11 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles((theme) => ({
-  smallScreenItem: {
-    [theme.breakpoints.up('md')]: {
-      display: 'none',
-    },
-  },
-}));
-
-export default useStyles;
