+++ client/src/components/NavBar/NavBar.tsx
@@ -1,30 +1,20 @@
-import { AppBar, Box, Typography } from '@material-ui/core';
-import logo from '../../Images/pawlogo.png';
-import LoggedInNavBarLinks from './NavBarLinks/loggedIn/NavBarLinks';
-import LoggedOutNavBarLinks from './NavBarLinks/loggedOut/NavBarLinks';
+import { AppBar, Box } from '@material-ui/core';
+import logo from '../../Images/logo.png';
+import NavBarLinks from './NavBarLinks/NavBarLinks';
 import useStyles from './useStyles';
 import CssBaseline from '@material-ui/core/CssBaseline';
-import { useAuth } from '../../context/useAuthContext';
+import NavBarMenu from './NavBarMenu/NavBarMenu';
 
 function NavBar(): JSX.Element {
   const classes = useStyles();
-  const user = useAuth().loggedInUser;
 
   return (
     <AppBar color="transparent" position="static" className={classes.root} elevation={0}>
       <CssBaseline />
       <Box className={classes.navWrapper}>
-        <Box className={classes.logoWrapper}>
-          <img src={logo} alt="logo" className={classes.logo} />
-          <Typography className={classes.logoText}>LovingSitter.</Typography>
-        </Box>
-        {!user ? (
-          <div>
-            <LoggedOutNavBarLinks />
-          </div>
-        ) : (
-          <LoggedInNavBarLinks user={user} />
-        )}
+        <img src={logo} alt="logo" className={classes.logo} />
+        <NavBarLinks />
+        <NavBarMenu />
       </Box>
     </AppBar>
   );
