+++ client/src/pages/Dashboard/Dashboard.tsx
@@ -1,12 +1,16 @@
 import Grid from '@material-ui/core/Grid';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import CircularProgress from '@material-ui/core/CircularProgress';
+import useStyles from './useStyles';
 import { useAuth } from '../../context/useAuthContext';
 import { useSocket } from '../../context/useSocketContext';
 import { useHistory } from 'react-router-dom';
+import ChatSideBanner from '../../components/ChatSideBanner/ChatSideBanner';
 import { useEffect } from 'react';
 
 export default function Dashboard(): JSX.Element {
+  const classes = useStyles();
+
   const { loggedInUser } = useAuth();
   const { initSocket } = useSocket();
 
@@ -24,10 +28,10 @@ export default function Dashboard(): JSX.Element {
   }
 
   return (
-    <Grid container component="main">
+    <Grid container component="main" className={`${classes.root} ${classes.dashboard}`}>
       <CssBaseline />
-      <Grid item>
-        <h1>MAIN DASHBOARD</h1>
+      <Grid item className={classes.drawerWrapper}>
+        <ChatSideBanner loggedInUser={loggedInUser} />
       </Grid>
     </Grid>
   );
