--- client/src/pages/Dashboard/profile/Profile.tsx
@@ -1,34 +1,0 @@
-import Grid from '@material-ui/core/Grid';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import CircularProgress from '@material-ui/core/CircularProgress';
-import { useAuth } from '../../../context/useAuthContext';
-import { useSocket } from '../../../context/useSocketContext';
-import { useHistory } from 'react-router-dom';
-import { useEffect } from 'react';
-
-export default function Profile(): JSX.Element {
-  const { loggedInUser } = useAuth();
-  const { initSocket } = useSocket();
-
-  const history = useHistory();
-
-  useEffect(() => {
-    initSocket();
-  }, [initSocket]);
-
-  if (loggedInUser === undefined) return <CircularProgress />;
-  if (!loggedInUser) {
-    history.push('/login');
-    // loading for a split seconds until history.push works
-    return <CircularProgress />;
-  }
-
-  return (
-    <Grid container component="main">
-      <CssBaseline />
-      <Grid item>
-        <h2>Profile</h2>
-      </Grid>
-    </Grid>
-  );
-}
